requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comtypes==1.2.0
distro==1.8.0
docx==0.2.4
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2023.12.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
Jinja2==3.1.2
llvmlite==0.41.1
lxml==4.9.3
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.3.7
openai-whisper==20231117
Pillow==10.1.0
psutil==5.9.6
pulsectl==23.5.2
PyAudio==0.2.14
pycaw==20230407
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
python-docx==1.1.0
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SoundCard==0.4.2
sounddevice==0.4.6
sympy==1.12
tiktoken==0.5.2
tk==0.1.0
torch==2.1.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.1.0
